[fix][cms] useless history/trash creation on merging branch into master (#3509)
* [add] reproducible spec

* [add] more specs about history/trashes

* [fix] cause error while destroying history/trashes in some cases

* [modify] improve efficiency by `constantize`

* [add] "#skip_histroy_trash" to skip creating history/trash

* [modify] stop creating thumb_file's history/trash to improve efficiencies

* [fix] stop creating history/trash record while merging branch into master

* [add] more specs about branch merge

* [modify] refactor w/ Rubocop

+++ app/models/cms/column/value/free.rb
@@ -87,6 +87,16 @@ class Cms::Column::Value::Free < Cms::Column::Value::Base
   end
 
   def destroy_files
-    files.destroy_all
+    if !_parent.respond_to?(:skip_history_trash)
+      files.destroy_all
+      return
+    end
+
+    file_ids.each_slice(20) do |ids|
+      SS::File.in(id: ids).to_a.map(&:becomes_with_model).each do |file|
+        file.skip_history_trash = _parent.skip_history_trash if file.respond_to?(:skip_history_trash)
+        file.destroy
+      end
+    end
   end
 end
